refactor: Achieve default of all repos in dir via standard options
Rather than having a bunch of custom code to handle the default case
where we wanted running "jira" without the --repo flag to look for
all git repos in the current directory, or the current directory
if itself was a git repo.

Now that same functionality is supplied by passing "./" and "./*"
as the default value of --repo.